[chore]DB config for travis

+++ app/views/bucketlist_item.py
@@ -1,74 +1,60 @@
 # /app/views/bucketlist_item.py
+from flask.views import MethodView
 from flask import request, jsonify, abort, make_response
 
 import decorator
 from app.models import BucketList, BucketListItem
-from . import bucketlist_item_blueprint as app
+from . import bucketlist_item_blueprint
 
 
-@app.route("/bucketlists/<int:id>/items/", methods=["POST"])
-@decorator.login_required
-def bucketlist_item(id, **kwargs):
-    user_id = kwargs["user_id"]
-    bucketlist = BucketList.query.filter_by(
-        id=id, created_by=user_id).first()
+class BucketListItemView(MethodView):
 
-    if not bucketlist:
-        # Raise an HTTPException with a 404 not found status code
-        abort(404)
+    @decorator.login_required
+    def post(self, id, user_id):
+        # user_id = kwargs["user_id"]
+        bucketlist = BucketList.query.filter_by(
+            id=id, created_by=user_id).first()
 
-    else:
-        name = str(request.data.get("name", ""))
-        if name:
-            item = BucketListItem(name=name, bucketlist_id=id)
-            item.save()
-            response = jsonify({
-                "id": item.id,
-                "name": item.name,
-                "date_created": item.date_created,
-                "date_modified": item.date_modified,
-                "bucketlist_id": id,
-                "done": item.done
-            })
-            return make_response(response), 201
-
-
-@app.route("/bucketlists/<int:id>/items/<int:item_id>",
-           methods=["PUT", "GET", "DELETE"])
-@decorator.login_required
-def bucketlist_item_manipulation(id, item_id, **kwargs):
-    user_id = kwargs["user_id"]
-    bucketlist = BucketList.query.filter_by(
-        id=id, created_by=user_id).first()
-    if not bucketlist:
-        # Raise an HTTPException with a 404 not found status code
-        abort(404)
-    else:
-        item = BucketListItem.query.filter_by(id=item_id).first()
-        if not item:
-            # Raise an HTTPException
+        if not bucketlist:
+            # Raise an HTTPException with a 404 not found status code
             abort(404)
 
-        if request.method == "PUT":
-            # Update bucketlist with new name
+        else:
             name = str(request.data.get("name", ""))
             if name:
-                item.name = name
+                item = BucketListItem(name=name, bucketlist_id=id)
                 item.save()
-                response = {
+                response = jsonify({
                     "id": item.id,
                     "name": item.name,
                     "date_created": item.date_created,
                     "date_modified": item.date_modified,
-                    "bucketlist_id": item.bucketlist_id
-                }
-                return make_response(jsonify(response)), 200
-            else:
-                response = {"message": "Enter a Valid Name"}
-                return make_response(jsonify(response)), 400
+                    "bucketlist_id": id,
+                    "done": item.done
+                })
+                return make_response(response), 201
 
-        elif request.method == "GET":
-            # Fetch the specified bucketlist
+
+class BucketListItemManipulationView(MethodView):
+
+    @decorator.login_required
+    def put(self, id, item_id, user_id):
+        bucketlist = BucketList.query.filter_by(
+            id=id, created_by=user_id).first()
+        if not bucketlist:
+            # Raise an HTTPException with a 404 not found status code
+            abort(404)
+        else:
+            item = BucketListItem.query.filter_by(id=item_id).first()
+        if not item:
+            # Raise an HTTPException
+            abort(404)
+
+        # Update bucketlist with new name
+        name = str(request.data.get("name", ""))
+        if name:
+            item.name = name
+            item.save()
             response = {
                 "id": item.id,
                 "name": item.name,
@@ -77,9 +63,62 @@ def bucketlist_item_manipulation(id, item_id, **kwargs):
                 "bucketlist_id": item.bucketlist_id
             }
             return make_response(jsonify(response)), 200
-
         else:
-            # delete the item using delete method
-            item.delete()
-            return {"message": "item {} deleted".
-                    format(item.id)}, 200
+            response = {"message": "Enter a Valid Name"}
+            return make_response(jsonify(response)), 400
+
+    @decorator.login_required
+    def get(self, id, item_id, user_id):
+        # Fetch the specified bucketlist
+        bucketlist = BucketList.query.filter_by(
+            id=id, created_by=user_id).first()
+        if not bucketlist:
+            # Raise an HTTPException with a 404 not found status code
+            abort(404)
+        else:
+            item = BucketListItem.query.filter_by(id=item_id).first()
+        if not item:
+            # Raise an HTTPException
+            abort(404)
+
+        response = {
+            "id": item.id,
+            "name": item.name,
+            "date_created": item.date_created,
+            "date_modified": item.date_modified,
+            "bucketlist_id": item.bucketlist_id
+        }
+        return make_response(jsonify(response)), 200
+
+    @decorator.login_required
+    def delete(self, id, item_id, user_id):
+        bucketlist = BucketList.query.filter_by(
+            id=id, created_by=user_id).first()
+        if not bucketlist:
+            # Raise an HTTPException with a 404 not found status code
+            abort(404)
+        else:
+            item = BucketListItem.query.filter_by(id=item_id).first()
+        if not item:
+            # Raise an HTTPException
+            abort(404)
+
+        # delete the item using delete method
+        item.delete()
+        return {"message": "item {} deleted".
+                format(item.id)}, 200
+
+
+# API resource
+bucketlist_item_view = BucketListItemView.as_view("bucketlist_item_view")
+item_manipulation_view = BucketListItemManipulationView.as_view(
+    "item_manipulation_view")
+
+# Rule for bucketlist with blueprint
+bucketlist_item_blueprint.add_url_rule("/bucketlists/<int:id>/items/",
+                                       view_func=bucketlist_item_view,
+                                       methods=["POST"])
+bucketlist_item_blueprint.add_url_rule(
+    "/bucketlists/<int:id>/items/<int:item_id>",
+    view_func=item_manipulation_view,
+    methods=["PUT", "GET", "DELETE"])
